requirements.txt
Pillow==8.4.0
Flask==2.1.1
tk==8.6.12
